refactor(codemode): simplify OpenAPI adapter split
